move old api in v1/; rename restapi.py to api.py
