+++ counterparty-lib/counterpartylib/test/estimate_fee_per_kb_test.py
@@ -4,7 +4,8 @@ import tempfile
 
 import bitcoin as bitcoinlib
 
-from counterpartylib.lib import api, backend, transaction
+from counterpartylib.lib import (backend, transaction)
+from counterpartylib.lib.v1 import api
 from counterpartylib.test import (
     util_test,
 )
